// fe/src/lib/api/generated/model/siblingPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { SexValue } from './sexValue';

export interface SiblingPayload {
  age?: number | null;
  educational_attainment?: string;
  is_self?: boolean;
  name?: string;
  occupation?: string;
  sex?: SexValue | null;
  /** @minimum 0 */
  sort_order: number;
}